Move more classes to their own file
