+++ Xamarin.PropertyEditing.Mac/Controls/Custom/SolidColorBrushEditor.cs
@@ -12,333 +12,6 @@ using Xamarin.PropertyEditing.ViewModels;
 
 namespace Xamarin.PropertyEditing.Mac
 {
-	abstract class ColorEditorView : NSView, INotifyingListner<SolidBrushViewModel>
-	{
-		protected const float padding = 3;
-		protected NotifyingViewAdaptor<SolidBrushViewModel> adaptor { get; }
-
-		public ColorEditorView (IntPtr handle) : base (handle)
-		{
-			adaptor = new NotifyingViewAdaptor<SolidBrushViewModel> (this);
-		}
-
-		[Export ("initWithCoder:")]
-		public ColorEditorView (NSCoder coder) : base (coder)
-		{
-			adaptor = new NotifyingViewAdaptor<SolidBrushViewModel> (this);
-		}
-
-		public ColorEditorView (CGRect frame) : base (frame)
-		{
-			adaptor = new NotifyingViewAdaptor<SolidBrushViewModel> (this);
-		}
-
-		public ColorEditorView () : base ()
-		{
-			adaptor = new NotifyingViewAdaptor<SolidBrushViewModel> (this);
-		}
-
-		public SolidBrushViewModel ViewModel {
-			get => adaptor.ViewModel;
-			set => adaptor.ViewModel = value;
-		}
-
-		protected virtual void OnViewModelChanged (SolidBrushViewModel oldModel)
-		{
-			OnPropertyChanged (ViewModel, new PropertyChangedEventArgs (nameof (SolidBrushViewModel.Color)));
-		}
-
-		protected virtual void OnPropertyChanged (object sender, PropertyChangedEventArgs e)
-		{
-		}
-
-		public override void MouseDragged (NSEvent theEvent)
-		{
-			//base.MouseDragged (theEvent);
-			UpdateFromEvent (theEvent);
-		}
-
-		public override void MouseDown (NSEvent theEvent)
-		{
-			//base.MouseDown (theEvent);
-			UpdateFromEvent (theEvent);
-		}
-
-		public virtual void UpdateFromEvent (NSEvent theEvent)
-		{
-		}
-
-		protected override void Dispose (bool disposing)
-		{
-			base.Dispose (disposing);
-
-			if (!disposing)
-				return;
-
-			adaptor.Disconnect ();
-		}
-
-		void INotifyingListner<SolidBrushViewModel>.OnViewModelChanged (SolidBrushViewModel oldModel)
-		{
-			OnViewModelChanged (oldModel);
-		}
-
-		void INotifyingListner<SolidBrushViewModel>.OnPropertyChanged (object sender, PropertyChangedEventArgs e)
-		{
-			OnPropertyChanged (sender, e);
-		}
-	}
-
-	abstract class ColorEditorLayer : CALayer
-	{
-		public ColorEditorLayer ()
-		{
-		}
-
-		public ColorEditorLayer (IntPtr handle) : base (handle)
-		{
-		}
-
-		abstract public void UpdateFromModel (EditorInteraction viewModel);
-		abstract public void UpdateFromLocation (EditorInteraction viewModel, CGPoint location);
-	}
-
-	class HistoryLayer : ColorEditorLayer
-	{
-		const float Margin = 3;
-		const float BorderRadius = 3;
-
-		public HistoryLayer ()
-		{
-			Clip.AddSublayer (Previous);
-			Clip.AddSublayer (Current);
-			AddSublayer (Clip);
-		}
-
-		public HistoryLayer (IntPtr handle) : base (handle)
-		{
-		}
-
-		readonly CALayer Previous = new CALayer ();
-		readonly CALayer Current = new CALayer ();
-		readonly CALayer Clip = new CALayer () {
-			BorderWidth = 1,
-			CornerRadius = BorderRadius,
-			BorderColor = new CGColor (.5f, .5f, .5f, .5f),
-			MasksToBounds = true,
-		};
-
-		public override void LayoutSublayers ()
-		{
-			base.LayoutSublayers ();
-
-			Clip.Frame = new CGRect (
-				Margin,
-				Margin,
-				Frame.Width - 2 * Margin,
-				Frame.Height - 2 * Margin);
-			
-			Clip.Contents = DrawingExtensions.GenerateCheckerboard (Clip.Frame);
-			var width = Clip.Frame.Width / 2;
-
-			Previous.Frame = new CGRect (0, 0, width, Clip.Frame.Height);
-			Current.Frame = new CGRect (width, 0, width, Clip.Frame.Height);
-		}
-
-		public override void UpdateFromModel (EditorInteraction interaction)
-		{
-			LayoutIfNeeded ();
-			Current.BackgroundColor = interaction.Color.ToCGColor ();
-			Previous.BackgroundColor = interaction.LastColor.ToCGColor ();
-		}
-
-		public override void UpdateFromLocation (EditorInteraction interaction, CGPoint location)
-		{
-			if (Previous == HitTest (location))
-				interaction.Color = interaction.LastColor;
-		}
-	}
-
-	class ShadeLayer : ColorEditorLayer
-	{
-		const float GripRadius = 4;
-		const float BorderRadius = 3;
-		const float Margin = 3;
-		ChannelEditor saturationEditor = new HsbSaturationChannelEditor ();
-		ChannelEditor brightnessEditor = new HsbBrightnessChannelEditor ();
-
-		public ShadeLayer ()
-		{
-			AddSublayer (Saturation);
-			Saturation.AddSublayer (Brightness);
-			AddSublayer (Grip);
-			float innerRadius = GripRadius - 1;
-
-			Grip.AddSublayer (new CALayer {
-				BorderWidth = 1,
-				BorderColor = new CGColor (0, 0, 0),
-				Frame = new CGRect (
-					GripRadius - innerRadius,
-					GripRadius - innerRadius,
-					innerRadius * 2,
-					innerRadius * 2),
-				CornerRadius = innerRadius
-			});
-		}
-
-		public ShadeLayer (IntPtr handle) : base (handle)
-		{
-		}
-
-		CALayer Grip = new CALayer {
-			BorderColor = new CGColor (1,1,1),
-			BorderWidth = 1,
-			CornerRadius = GripRadius,
-		};
-
-		CAGradientLayer Brightness = new CAGradientLayer {
-			Colors = new[] {
-					new CGColor (0f, 0f, 0f, 1f),
-					new CGColor (0f, 0f, 0f, 0f)
-				},
-			CornerRadius = BorderRadius,
-		};
-
-		CAGradientLayer Saturation = new CAGradientLayer {
-			Colors = new[] {
-					new CGColor (1f, 1f, 1f),
-					new CGColor (1f, .3f, 0f)
-				},
-			BackgroundColor = NSColor.Black.CGColor,
-			BorderColor = new CGColor (.5f, .5f, .5f, .5f),
-			BorderWidth = 1,
-			CornerRadius = BorderRadius,
-		};
-
-        public override void LayoutSublayers ()
-        {
-            base.LayoutSublayers();
-
-			Saturation.Frame = new CGRect (Margin, Margin, Frame.Width - 2 * Margin, Frame.Height - 2 * Margin);
-			Brightness.Frame = Saturation.Bounds;
-			Saturation.StartPoint = new CGPoint (0, .5);
-			Saturation.EndPoint = new CGPoint (1, .5);
-        }
-
-		CommonColor c;
-		public override void UpdateFromModel (EditorInteraction interaction)
-		{
-			LayoutIfNeeded ();
-			var color = interaction.Color;
-
-			var sat = saturationEditor.LocationFromColor (Saturation, color);
-			var bright = brightnessEditor.LocationFromColor (Brightness, color);
-
-			var x = sat.X;
-			var y = bright.Y + Saturation.Frame.Y;
-
-			Grip.Frame = new CGRect (x - GripRadius, y - GripRadius, GripRadius * 2, GripRadius * 2);
-
-			if (interaction.StartColor.ToCGColor () != Saturation.Colors.Last())
-				saturationEditor.UpdateGradientLayer (Saturation, interaction.StartColor.HueColor);
-		}
-
-		public override void UpdateFromLocation (EditorInteraction interaction, CGPoint location)
-		{
-			var loc = location;
-			var frame = Saturation.Frame;
-
-			if (interaction.ViewModel == null)
-				return;
-
-			var color = interaction.Color;
-			var saturation = saturationEditor.ValueFromLocation (Saturation, loc);
-			var brightness = saturationEditor.ValueFromLocation (
-				Brightness, 
-			    new CGPoint (loc.X + Brightness.Frame.X, loc.Y + Brightness.Frame.Y));
-			
-			c = interaction.Color = interaction.Color.UpdateHSB (saturation: saturation, brightness: brightness);
-		}
-    }
-
-	class HueLayer : ColorEditorLayer
-	{
-		const float Margin = 3;
-		const float BorderRadius = 3;
-		const float GripRadius = 3;
-
-		ChannelEditor hueEditor = new HsbHueChannelEditor ();
-
-		public CGColor GripColor
-		{
-			get => Grip.BorderColor;
-			set => Grip.BorderColor = value;
-		}
-
-		public HueLayer ()
-		{
-			hueEditor.UpdateGradientLayer (Colors, new CommonColor (0,255,0));
-			AddSublayer (Colors);
-			AddSublayer (Grip);
-		}
-
-		public HueLayer (IntPtr handle) : base (handle)
-		{
-		}
-
-		CAGradientLayer Colors = new CAGradientLayer {
-			BorderColor = new CGColor (.5f, .5f, .5f, .5f),
-			StartPoint = new CGPoint (0,1),
-			EndPoint = new CGPoint (0,0),
-			BorderWidth = 1,
-			CornerRadius = BorderRadius,
-		};
-
-		CALayer Grip = new CALayer {
-			BorderColor = NSColor.Text.CGColor,
-			BorderWidth = 2,
-			CornerRadius = GripRadius,
-		};
-
-		CommonColor c;
-		public override void UpdateFromModel (EditorInteraction interaction)
-		{
-			LayoutIfNeeded ();
-
-			var color = interaction.Color;
-			if (c == color)
-				return;
-
-			var loc = hueEditor.LocationFromColor (Colors, color);
-			Grip.Frame = new CGRect (1, loc.Y - Grip.Frame.Height / 2f, Grip.Frame.Width, Grip.Frame.Height);
-		}
-
-		public override void UpdateFromLocation (EditorInteraction interaction, CGPoint location)
-		{
-			var loc = location;
-			var clos = Math.Min (Colors.Frame.Height, Math.Max (0, loc.Y - Colors.Frame.Y));
-
-			Grip.Frame = new CGRect (1, clos + Colors.Frame.Y - Grip.Frame.Height / 2f, Frame.Width - 2, 2 * GripRadius);
-			var hue = (1 - clos/ Colors.Frame.Height) * 360;
-
-			if (interaction == null)
-				return;
-
-			var color = interaction.Color;
-			c = interaction.Color = hueEditor.UpdateColorFromLocation (
-				Colors,
-				interaction.Color,
-				loc);
-		}
-
-        public override void LayoutSublayers()
-        {
-			base.LayoutSublayers ();
-			Colors.Frame = Bounds.Inset (2, 2);
-			Grip.Frame = new CGRect (Grip.Frame.X, Grip.Frame.Y, Frame.Width - 2, 2 * GripRadius);
-		}
-	}
-
 	class EditorInteraction : NotifyingObject
 	{
 		public EditorInteraction (SolidBrushViewModel viewModel, ColorEditorLayer layer)
@@ -378,9 +51,9 @@ namespace Xamarin.PropertyEditing.Mac
 
 	class SolidColorBrushEditor : ColorEditorView
 	{
-		ShadeLayer Shade = new ShadeLayer ();
-		HueLayer Hue = new HueLayer ();
-		HistoryLayer History = new HistoryLayer ();
+		readonly ShadeLayer Shade = new ShadeLayer ();
+		readonly HueLayer Hue = new HueLayer ();
+		readonly HistoryLayer History = new HistoryLayer ();
 
 		readonly CALayer Background = new CALayer {
 			CornerRadius = 3,
